changes datatype for ProductPrice field and looks more into cart functionality.

+++ HayleesThreads/Models/Product.cs
@@ -14,7 +14,7 @@ namespace HayleesThreads.Models
 
     public string ProductName { get; set; }
   
-    public decimal ProductPrice { get; set; }
+    public float ProductPrice { get; set; }
 
     public string ProductImage { get; set; }
